Support enums used across packages (#10)
All of the types and enum names have to be fully-qualified with their
package name so that usages across packages can be identified.

Also cleaned up "-show-enums" into a multiline that is easier to read
(it's still not a porcelain format).

Go lets you specify a different package name in the source files from
the directory they reside in. Right now I haven't built a way to fetch
all the real package names from the directories so I'll just assume that
people would only remove punctuation (like a dash or underscore).